+++ test/skogsra/app_test.exs
@@ -14,7 +14,7 @@ defmodule Skogsra.AppTest do
     test "when it's defined, gets value" do
       ApplicationMock.put_env(:app, :key, 42)
 
-      env = Env.new(nil, :app, :key, [])
+      env = Env.new(nil, :app, :key, type: :integer)
 
       assert 42 == App.get_env(env)
     end
@@ -22,13 +22,21 @@ defmodule Skogsra.AppTest do
     test "when it's defined in a namespace, gets value" do
       ApplicationMock.put_env(:app, My.Custom.Namespace, key: 42)
 
-      env = Env.new(My.Custom.Namespace, :app, :key, [])
+      env = Env.new(My.Custom.Namespace, :app, :key, type: :integer)
 
       assert 42 == App.get_env(env)
     end
 
     test "when it's not defined, does not get the value" do
       env = Env.new(nil, :app, [:a, :b], [])
+
+      assert nil == App.get_env(env)
+    end
+
+    test "when cast is not possible, does not get the value" do
+      ApplicationMock.put_env(:app, :key, 42.0)
+
+      env = Env.new(nil, :app, :key, type: :integer)
 
       assert nil == App.get_env(env)
     end